+++ src/main/java/com/yuuko/core/commands/moderation/commands/NukeCommand.java
@@ -33,17 +33,17 @@ public class NukeCommand extends Command {
         }
 
         if(!Sanitiser.isNumber(e.getCommand().get(1))) {
-            EmbedBuilder embed = new EmbedBuilder().setTitle("Invalid Input").setDescription("Input must be a positive integer between **2** and **99** or a channel, e.g. #general.");
+            EmbedBuilder embed = new EmbedBuilder().setTitle("Invalid Input").setDescription("Input must be a positive integer between **2** and **100** or a channel, e.g. #general.");
             MessageHandler.sendMessage(e, embed.build());
             return;
         }
 
-        final int value = Integer.parseInt(e.getCommand().get(1));
+        int value = Integer.parseInt(e.getCommand().get(1));
 
-        if(value < 2 || value > 99) {
-            EmbedBuilder embed = new EmbedBuilder().setTitle("Invalid Input").setDescription("Input must be a positive integer between **2** and **99** or a channel, e.g. #general.");
-            MessageHandler.sendMessage(e, embed.build());
-            return;
+        if(value < 2) {
+            value = 2;
+        } else if(value > 99) {
+            value = 99;
         }
 
         // Filter out old messages from the mass delete list.